+++ CHANGELOG.md
@@ -11,6 +11,16 @@ Check [RELEASE](https://github.com/we-mak/w-design/releases)
 
 --
 
+### v0.48.1
+
+- **Layout**: Base layout for App
+- **Image**
+  - Add loading props default to `lazy`
+  - Use `width`, `height` attribute support for prevent content layout shift (CLS)
+  - Fixed https://github.com/we-mak/w-design/issues/203
+
+--
+
 ### v0.47.0
 
 - ✅ #144 **Steps**